Prototyping

+++ src/wm/window.h
@@ -2,10 +2,12 @@
 
 #include <xcb/xcb.h>
 
+#include "group.h"
+
 typedef struct {
     xcb_window_t id;
     xcb_window_t parent;
-    unsigned int x, y, height, width;
+    float x, y, height, width; /* X and Y are relative to the group */
 } window_t;
 
 window_t *manage_window(xcb_window_t);
